+++ src/main/java/com/standard/repository/ItensTipoMedidaRepository.java
@@ -3,7 +3,7 @@ package com.standard.repository;
 import com.standard.entity.ItensTipoMedidaEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ItensTipoMedidaRepository extends JpaRepository<ItensTipoMedidaEntity, Integer> {
+public interface ItensTipoMedidaRepository extends JpaRepository<ItensTipoMedidaEntity, Long> {
 
     ItensTipoMedidaEntity findByMedidaCodigo(Integer medida_codigo);
